requirements.txt
apify-client
fastapi
uvicorn[standard]
requests
pydantic
python-multipart
dotenv
google-cloud-aiplatform
asyncpg
psycopg2-binary
redis
google-generativeai